# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "EvacConnect API Documentation",
    "version": "1.0.0",
    "title": "EvacConnect API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ec2-52-64-253-171.ap-southeast-2.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/person-info":{
      "get": {
        "tags": ["users"],
        "summary": "Get person info",
        "produces": ["application/json"],
        "parameters":[
          {
            "in" : "header",
            "name" : "Authorization",
            "type" : "string",
            "required": true
          },
          {
            "in" : "query",
            "name" : "userId",
            "type" : "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Request successful",
            "schema": {
              "$ref": "#/definitions/ApiPersonStructure"
            }
          },
          "400": {
            "description": "Invalid internal operation(SQL)",
            "schema": {
              "$ref": "#/definitions/ApiResponseStructure"
            }
          }
        }
      }
    },
    "/update-person-info":{
      "patch":{
        "summary": "Update person information",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "in" : "header",
            "name" : "Authorization",
            "type" : "string",
            "required": true
          },
          {
            "in": "body",
            "name": "person information",
            "description": "Information about the user to be updated",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "type":"string",
                  "description": "The user's first name"
                },
                "last_name": {
                  "type":"string",
                  "description": "The user's family name"
                },
                "email": {
                  "type": "string",
                  "description": "The user's validated email used  in signing up"
                },
                "phone_number": {
                  "type": "string",
                  "description": "Landline contact number"
                },
                "occupation": {
                  "type": "string",
                  "description": "The user's occupation"
                }
              },
              "required": [
                "first_name",
                "last_name",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person Info Updated Successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }

      }
    },
...
  },
  "definitions": {
    "ApiResponseStructure": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "ApiPersonStructure": {
      "type": "object",
      "properties": {
        "first_name": { "type":"string", "required": true },
        "last_name": { "type":"string", "required": true },
        "email": { "type": "string", "required": true },
        "phone_number": { "type": "string" },
        "account_name": { "type": "string" },
        "occupation": { "type": "string"}
      }
    }
  }
}